// src/protocol/types.ts
export interface User {
  AccountID: number;
  Avatar?: string;
  AwaySince?: string;
  BanMute: boolean;
  BanVotes: boolean;
  BanSpecChat: boolean;
  BattleID?: number;
  Clan?: string;
  Country?: string;
  DisplayName?: string;
  Faction?: string;
  InGameSince?: string;
  IsAdmin: boolean;
  IsBot: boolean;
  LobbyVersion?: string;
  Name?: string;
  SteamID?: string;
  Badges?: string[];
  Icon?: string;
  EffectiveMmElo: number;
  EffectiveElo: number;
  Level: number;
  Rank: number;
}

export interface UserDisconnected {
  Name?: string;
  Reason?: string;
}

export interface Say {
  IsEmote: boolean;
  Place: E.SayPlace;
  Ring: boolean;
  Target?: string;
  Text?: string;
  Time?: string;
  User?: string;
  Source?: E.SaySource;
}
// ...
export interface OnPartyInvite {
  PartyID: number;
  UserNames?: string[];
  TimeoutSeconds: number;
}